Complete release v0.9-a1 Support all unsigned operation. Done ucode cascade resolution protocol. Structure UcodeRP API with ref implementation of client and server.

+++ uid/src/com/dadfha/uid/UrpRecieve.java
@@ -46,7 +46,7 @@ public class UrpRecieve extends UrpPacket {
 	 * @return Error
 	 */
 	public Error getErrorCode() {
-		return Error.valueOf(getOperator());
+		return Error.valueOf( (short) getOperator() );
 	}
 	
 	/**
